+++ app/src/main/java/com/intive/patronage/smarthome/notifications/model/Notification.kt
@@ -3,7 +3,7 @@ package com.intive.patronage.smarthome.notifications.model
 import com.squareup.moshi.Json
 
 data class Notification(
-    @Json(name = "id") val id: Int,
+    @Json(name = "id") val id: Long,
     @Json(name = "timestamp") val timestamp: Long,
     @Json(name = "type") val type: String,
     @Json(name = "sensorId") val sensorId: Int,
